Huge refactoring, split classes by modules, moved all real message processing to GameManager
