+++ modules/sea_battle_package/__init__.py
@@ -1,4 +1,8 @@
+from .game_constants import *
+from .game_context import *
+from .game_manager import *
 from .ship_processing import *
-from .game_processing import *
+from .team import *
+from utils import *
 
 __version__ = (0, 1, 0)
