+++ utils.py
@@ -110,9 +110,3 @@ def rated_operation(operation, args):
                 raise
         except:
             raise
-
-def enum(*sequential, **named):
-    enums = dict(zip(sequential, range(len(sequential))), **named)
-    reverse = dict((value, key) for key, value in enums.iteritems())
-    enums['reverse_mapping'] = reverse
-    return type('Enum', (), enums)
